Rename to "pIqaD nap"
For some unknown reason, omitting "tlhIngan" from the beginning of the
font name allows the font to occupy a more preferable position in the
font cascade on iOS. With the "tlhIngan pIqaD nap" name, for some of
the letters, glyphs from other fonts that provide symbols within the
private use area designated for pIqaD in the Conscript Unicode Registry
would appear instead of pIqaD when the pIqaD font wasn't explicitly
selected, and with the "pIqaD nap" name, the glyphs from the pIqaD font
would take precedence.

My best guess for why this is the case is that the undesired glyphs
appear to come from the system-provided font "Symbol", and perhaps
the precedence in the font cascade is based on a case-insensitive
alphabetical sort.